requirements.txt
asgiref==3.8.1
botocore==1.35.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Django==5.1
django-rest-knox==5.0.1
djangorestframework==3.15.2
dynaconf==3.2.6
hvac==2.3.0
idna==3.7
loguru==0.7.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
validators==0.33.0
